[react] change over to new application
